requirements.txt
asgiref==3.8.1
boto3==1.34.126
botocore==1.34.126
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.2.0
Django==4.2.13
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
lxml==5.2.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.1
pillow==9.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-docx-replace==0.4.4
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.30.1
whitenoise==6.6.0
